requirements.txt
#설정 파일
PyYAML==6.0.2           #설정을 위한 YAML 패키지

#환경변수
python-dotenv==1.0.1           #환경변수 설정을 위한 패키지

# 웹 스크래핑 및 데이터 수집 관련 패키지
requests==2.32.3           # HTTP 요청 라이브러리
beautifulsoup4==4.12.3     # HTML 파싱 및 스크래핑 라이브러리
selenium==4.25.0           # 동적 웹페이지 스크래핑을 위한 라이브러리

# 데이터 처리 및 분석 패키지
pandas==2.2.3              # 데이터프레임을 사용한 데이터 분석

# 테스트 및 디버깅 패키지
pytest==8.3.3              # 파이썬 테스트 프레임워크
psycopg2~=2.9.10
Flask~=3.1.0
bs4~=0.0.2
sentence-transformers~=3.3.1
scikit-learn~=1.6.0
konlpy~=0.6.0